+++ website/sidebars.js
@@ -19,16 +19,33 @@ const sidebars = {
       type: 'category',
       collapsed: false,
       label: 'Getting Started',
-      items: ['installation', 'configuration', 'first-composite', 'client-setup'],
+      link: { type: 'doc',id: 'getting-started'},
+      items: [
+        { type: 'doc', id: 'set-up-your-environment', label: 'Set up your environment' },
+        { type: 'doc', id: 'create-your-composite', label: 'Create your composite' },
+        { type: 'doc', id: 'interact-with-data', label: 'Interact with data' },
+        { type: 'doc', id: 'next-steps', label: 'Next Steps' },
+      ],
     },
+    { 
+      type: 'category',
+      collapsed: true,
+      label: 'Core Concepts',
+      link: { type: 'doc',id: 'core-concepts'},
+      items: [
+        { type: 'doc', id: 'graph-structure', label: 'Graph' },
+        { type: 'doc', id: 'data-modeling-concepts', label: 'Models' },
+        { type: 'doc', id: 'database', label: 'Database' },
+      ]
+    },
+    { type: 'doc', id: 'community', label: 'Community' },
     {
       type: 'category',
       collapsed: false,
       label: 'Guides',
       link: { type: 'generated-index' },
       items: [
-        'guides/concepts-overview',
-        'guides/data-composition',
+        "guides/data-composition",
         {
           type: 'category',
           label: 'Using composites',
@@ -54,17 +71,31 @@ const sidebars = {
           ],
         },
         {
-          type: 'category',
-          label: 'Interacting with data',
-          link: { type: 'generated-index' },
-          items: [
-            { type: 'doc', id: 'guides/interacting/queries' },
-            { type: 'doc', id: 'guides/interacting/mutations' },
-            { type: 'doc', id: 'guides/interacting/using-apollo' },
-            { type: 'doc', id: 'guides/interacting/using-relay' },
-          ],
-        },
-      ],
+          "type": "category",
+          "label": "Interacting with data",
+          "link": {
+            "type": "generated-index"
+          },
+          "items": [
+            {
+              "type": "doc",
+              "id": "guides/interacting/queries"
+            },
+            {
+              "type": "doc",
+              "id": "guides/interacting/mutations"
+            },
+            {
+              "type": "doc",
+              "id": "guides/interacting/using-apollo"
+            },
+            {
+              "type": "doc",
+              "id": "guides/interacting/using-relay"
+            }
+          ]
+        }
+      ]
     },
   ],
   api: [